Handling the return objects from the backend API
Following methodes of the auth_service were updated:
- login()
- register()
- getAuthenticationCredentials() [Please pay attention to the token
validation process. Shall only the access token be validated? Frontend
team shall discuss this in person]
- validateToken() [The implementation sahll also be further discussed in
person]

A new enum TokenType was introduced.

As a suggestion: Wrap all tokens in a class to allow the easier
accessibility to expiration times for example.